+++ apps/website/app/api/embeddings/route.ts
@@ -1,16 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
-import OpenAI from "openai";
 import cors from "~/utils/llm/cors";
-
-const apiKey = process.env.OPENAI_API_KEY;
-
-if (!apiKey) {
-  console.error(
-    "Missing OPENAI_API_KEY environment variable. The embeddings API will not function.",
-  );
-}
-
-const openai = apiKey ? new OpenAI({ apiKey }) : null;
+import { genericEmbedding } from "~/utils/supabase/apiUtils";
 
 type RequestBody = {
   input: string | string[];
@@ -20,60 +10,8 @@ type RequestBody = {
   encoding_format?: "float" | "base64";
 };
 
-const OPENAI_REQUEST_TIMEOUT_MS = 30000;
-
-async function openai_embedding(
-  input: string | string[],
-  model: string,
-  dimensions?: number,
-): Promise<number[] | number[][] | undefined> {
-  let options: OpenAI.EmbeddingCreateParams = {
-    model,
-    input,
-  };
-  if (dimensions) {
-    options = { ...options, ...{ dimensions } };
-  }
-
-  const embeddingsPromise = openai!.embeddings.create(options);
-  const timeoutPromise = new Promise<never>((_, reject) =>
-    setTimeout(
-      () => reject(new Error("OpenAI API request timeout")),
-      OPENAI_REQUEST_TIMEOUT_MS,
-    ),
-  );
-
-  const response = await Promise.race([embeddingsPromise, timeoutPromise]);
-  const embeddings = response.data.map((d) => d.embedding);
-  if (Array.isArray(input)) return embeddings;
-  else return embeddings[0];
-}
-
-export async function generic_embedding(
-  input: string | string[],
-  model: string,
-  provider: string,
-  dimensions?: number,
-): Promise<number[] | number[][] | undefined> {
-  provider = provider || "openai";
-  if (provider == "openai") {
-    return await openai_embedding(input, model, dimensions);
-  }
-}
-
 export const POST = async (req: NextRequest): Promise<NextResponse> => {
   let response: NextResponse;
-
-  if (!apiKey) {
-    response = NextResponse.json(
-      {
-        error: "Server configuration error.",
-        details: "Embeddings service is not configured.",
-      },
-      { status: 500 },
-    );
-    return cors(req, response) as NextResponse;
-  }
 
   try {
     const body: RequestBody = await req.json();
@@ -92,7 +30,7 @@ export const POST = async (req: NextRequest): Promise<NextResponse> => {
       return cors(req, response) as NextResponse;
     }
 
-    const embeddings = await generic_embedding(
+    const embeddings = await genericEmbedding(
       input,
       model,
       provider,
